third_party/WebKit/LayoutTests/imported/web-platform-tests/web-animations/animatable/animate-expected.txt

This is a testharness.js-based test.
PASS Element.animate() creates an Animation object 
PASS Element.animate() creates an Animation object with a KeyframeEffect 
FAIL Element.animate() accepts a property-indexed keyframe specification anim.effect.getFrames is not a function
FAIL Element.animate() accepts a frame-indexed keyframe specification anim.effect.getFrames is not a function
FAIL Element.animate() accepts a single-valued keyframe specification anim.effect.getFrames is not a function
PASS Element.animate() accepts a double as an options argument 
PASS Element.animate() accepts a KeyframeAnimationOptions argument 
PASS Element.animate() accepts an absent options argument 
PASS Element.animate() correctly sets the id attribute when no id is specified 
PASS Element.animate() correctly sets the id attribute 
PASS Element.animate() correctly sets the Animation's timeline 
PASS Element.animate() correctly sets the Animation's timeline when triggered on an element in a different document 
PASS Element.animate() calls play on the Animation 
FAIL CSSPseudoElement.animate() creates an Animation object document.getAnimations is not a function
FAIL CSSPseudoElement.animate() creates an Animation object targeting to the correct CSSPseudoElement object document.getAnimations is not a function
Harness: the test ran to completion.
